Move celery integration to own file
